feat: 커뮤니티, 숏폼 권한 수정

+++ src/main/java/MeoipZi/meoipzi/controller/CommunityController.java
@@ -5,20 +5,17 @@ import MeoipZi.meoipzi.domain.Community;
 import MeoipZi.meoipzi.dto.CommunityDto.CommunityListResponseDTO;
 import MeoipZi.meoipzi.dto.CommunityDto.CommunityRequestDTO;
 import MeoipZi.meoipzi.dto.CommunityDto.CommunityResponseDTO;
-import MeoipZi.meoipzi.dto.CommunityDto.CommunityUpdateRequestDTO;
 import MeoipZi.meoipzi.repository.CommunityRepository;
 import MeoipZi.meoipzi.service.CommunityService;
 import lombok.RequiredArgsConstructor;
-import org.apache.http.HttpStatus;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.parameters.P;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
-import java.io.IOException;
 import java.util.List;
-import java.util.Optional;
 
 @Controller
 @RequiredArgsConstructor
@@ -32,38 +29,52 @@ public class CommunityController {
 
     /* 커뮤니티 게시글 리스트 보기 */
     @GetMapping("")
-    public ResponseEntity<List<CommunityListResponseDTO>> getCommunityList(){
+    public ResponseEntity<?> getCommunityList(){
         try {
             List<CommunityListResponseDTO> communityListResponseDTOS = communityService.getCommunityList();
 
-            return ResponseEntity.ok(communityListResponseDTOS); // 조회된 커뮤니티 글 리스트 형식 반환
+            return new ResponseEntity<>(communityListResponseDTOS, HttpStatus.OK);
+            // 조회된 커뮤니티 글 리스트 형식 반환
         } catch(Exception e) {
-            return ResponseEntity.status(HttpStatus.SC_INTERNAL_SERVER_ERROR).build();
+            return new ResponseEntity<>("커뮤니티 게시글 목록 조회에 실패하였습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     /* 커뮤니티 게시글 하나 상세 조회 */
     @GetMapping("/{communityId}")
-    public ResponseEntity<CommunityResponseDTO> getOneCommunity(@PathVariable Long communityId) {
+    public ResponseEntity<?> getOneCommunity(@PathVariable Long communityId) {
         try {
             CommunityResponseDTO communityResponseDTO = communityService.viewCommunity(communityId);
 
-            return ResponseEntity.ok(communityResponseDTO);
+            return new ResponseEntity<>(communityResponseDTO, HttpStatus.OK);
         } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.SC_INTERNAL_SERVER_ERROR).build();
+            return new ResponseEntity<>("게시글 상세 조회에 실패하였습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
-    /* 게시글 등록 */
+    /* 게시글 등록 - 사용자 로그인 여부 필요 */
     @PostMapping("")
-    public void createPost(CommunityRequestDTO communityRequestDTO){
-        communityService.join(communityRequestDTO);
+    public ResponseEntity<?> createPost(CommunityRequestDTO communityRequestDTO){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        communityRequestDTO.setUserName(authentication.getName()); // 현재 로그인한 작성자의 이름으로 작성하기
+
+        if(authentication.isAuthenticated()) {
+            return communityService.saveCommunity(communityRequestDTO);
+        }
+        else {
+            return new ResponseEntity<>("로그인하지 않았으므로 게시글 작성 권한이 없습니다.", HttpStatus.FORBIDDEN);
+        }
     }
 
-    /* 게시글 수정 */
+/*
+    */
+/* 게시글 수정 - 사용자 로그인 여부 필요 *//*
+
     @PatchMapping("/{communityId}")
     public ResponseEntity<String> updatePost(@PathVariable Long communityId, CommunityUpdateRequestDTO communityUpdateRequestDTO) {
         try {
+            Authentication authentication = SecurityContextHolder.createEmptyContext().getAuthentication();
+
             Optional<Community> op = communityRepository.findById(communityId);
             if(op.isPresent()) {
                 Community originalCommunity = op.get();
@@ -82,16 +93,20 @@ public class CommunityController {
             return ResponseEntity.ofNullable("게시글 수정에 실패하였습니다.");
         }
     }
+*/
 
     /* 게시글 삭제 */
     @DeleteMapping("/{communityId}")
-    public ResponseEntity<String> deletePost(@PathVariable Long communityId) {
-        try {
-            communityService.deleteCommunity(communityId);
-            return ResponseEntity.ok("게시글 삭제에 성공하였습니다.");
-        } catch (Exception e) {
-            return ResponseEntity.ofNullable("게시글 삭제 중에 오류가 발생하였습니다.");
+    public ResponseEntity<?> deletePost(@PathVariable Long communityId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(authentication.isAuthenticated()) {
+            try {
+                communityService.deleteCommunity(communityId);
+                return new ResponseEntity<>("게시글 삭제에 성공하였습니다.", HttpStatus.OK);
+            } catch (Exception e) {
+                return new ResponseEntity<>("게시글 삭제에 실패하였습니다.", HttpStatus.INTERNAL_SERVER_ERROR);}
         }
+        else return new ResponseEntity<>("게시글 삭제 권한이 없습니다.", HttpStatus.FORBIDDEN);
     }
 
 }
